requirements.txt
Pillow
opencv-python
numpy
matplotlib
flake8
autopep8